requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
anyio==4.13.0
    # via watchfiles
asgiref==3.11.1
    # via django
click==8.3.1
    # via uvicorn
django==6.0.3
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via anyio
markdown==3.10.2
    # via -r requirements.in
python-dotenv==1.2.2
    # via uvicorn
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   uvicorn
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via anyio
uvicorn==0.42.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
